+++ package.json
@@ -30,14 +30,10 @@
     ],
     "nohoist": [
       "**/vite",
-      "**/vite/**",
-      "**/prisma",
-      "**/prisma/**",
-      "**/@prisma/client",
-      "**/@prisma/client/**"
+      "**/vite/**"
     ]
   },
-  "packageManager": "bun@1.1.3",
+  "packageManager": "bun@1.1.9",
   "dependencies": {
     "-": "^0.0.1",
     "@datocms/cma-client-node": "^3.3.1",
